% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datbinary.R
\docType{data}
\name{datbinary}
\alias{datbinary}
\title{Simulated dataset, binary outcome}
\format{
An object of class \code{data.frame} with 1000 rows and 13 columns.
}
\usage{
data(datbinary)
}
\description{
Simulated dataset, binary outcome
}
\examples{
data(datbinary)
head(datbinary)
}
\keyword{datasets}
