% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datcont.R
\docType{data}
\name{datcont}
\alias{datcont}
\title{Simulated dataset, continuous outcome}
\format{
An object of class \code{data.frame} with 500 rows and 11 columns.
}
\usage{
data(datcont)
}
\description{
Simulated dataset, continuous outcome
}
\examples{
data(datcont)
head(datcont)
}
\keyword{datasets}
