% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbinary.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(x)
}
\arguments{
\item{x}{A real number between 0 and 1}
}
\value{
log(x/(1-x))
}
\description{
Calculates the logit of a real number between 0 and 1
}
\examples{
logit(0.2)
}
